+++ emanote/src/Emanote/Model/Type.hs
@@ -19,6 +19,7 @@ import Emanote.Model.Link.Rel qualified as Rel
 import Emanote.Model.Note (
   IxNote,
   Note,
+  ParseContext,
   noteHasFeed,
  )
 import Emanote.Model.Note qualified as N
@@ -43,7 +44,6 @@ import Optics.Core (Prism')
 import Optics.Operators ((%~), (.~), (^.))
 import Optics.TH (makeLenses)
 import Relude
-import Text.Pandoc.Scripting (ScriptingEngine)
 
 data Status = Status_Loading | Status_Ready
   deriving stock (Eq, Show)
@@ -55,8 +55,7 @@ data ModelT encF = Model
   , _modelRoutePrism :: encF (Prism' FilePath SiteRoute)
   , _modelPandocRenderers :: EmanotePandocRenderers Model LMLRoute
   -- ^ Dictates how exactly to render `Pandoc` to Heist nodes.
-  , _modelScriptingEngine :: Maybe ScriptingEngine
-  , _modelPluginBaseDirs :: [FilePath]
+  , _modelParseContext :: Maybe ParseContext
   , _modelCompileTailwind :: Bool
   , _modelInstanceID :: UUID
   -- ^ An unique ID for this process's model. ID changes across processes.
@@ -97,16 +96,15 @@ withRoutePrism enc Model {..} =
   let _modelRoutePrism = Identity enc
    in Model {..}
 
-emptyModel :: Set Loc -> Ema.CLI.Action -> EmanotePandocRenderers Model LMLRoute -> Maybe ScriptingEngine -> [FilePath] -> Bool -> UUID -> Stork.IndexVar -> ModelEma
-emptyModel layers act ren scriptingEngine pluginBaseDirs ctw instanceId storkVar =
+emptyModel :: Set Loc -> Ema.CLI.Action -> EmanotePandocRenderers Model LMLRoute -> Maybe ParseContext -> Bool -> UUID -> Stork.IndexVar -> ModelEma
+emptyModel layers act ren parseContext ctw instanceId storkVar =
   Model
     { _modelStatus = Status_Loading
     , _modelLayers = layers
     , _modelEmaCLIAction = act
     , _modelRoutePrism = Const ()
     , _modelPandocRenderers = ren
-    , _modelScriptingEngine = scriptingEngine
-    , _modelPluginBaseDirs = pluginBaseDirs
+    , _modelParseContext = parseContext
     , _modelCompileTailwind = ctw
     , _modelInstanceID = instanceId
     , -- Inject a placeholder `index.md` to account for the use case of emanote
@@ -130,19 +128,8 @@ inLiveServer :: Model -> Bool
 inLiveServer = Ema.CLI.isLiveServer . _modelEmaCLIAction
 
 modelInsertNote :: Note -> ModelT f -> ModelT f
-modelInsertNote note =
-  modelNotes
-    %~ ( Ix.updateIx r note
-          -- Insert folder placeholder automatically for ancestor paths
-          >>> injectAncestors (N.noteAncestors note)
-          >>> dropRedundantAncestor r
-       )
-      >>> modelRels
-    %~ updateIxMulti r (Rel.noteRels note)
-      >>> modelTasks
-    %~ updateIxMulti r (Task.noteTasks note)
-  where
-    r = note ^. N.noteRoute
+modelInsertNote =
+  modelInsertNotes . one
 
 modelInsertNotes :: [Note] -> ModelT f -> ModelT f
 modelInsertNotes [] =
@@ -226,18 +213,6 @@ modelDeleteNote k model =
       let folderR = R.withLmlRoute coerce k
       guard $ N.hasChildNotes folderR $ model ^. modelNotes
       pure folderR
-
--- | Like `Ix.updateIx`, but works for multiple items.
-updateIxMulti ::
-  (Ix.IsIndexOf ix ixs, Ix.Indexable ixs a) =>
-  ix ->
-  Ix.IxSet ixs a ->
-  Ix.IxSet ixs a ->
-  Ix.IxSet ixs a
-updateIxMulti r new rels =
-  let old = rels @= r
-      deleteMany = foldr Ix.delete
-   in new `Ix.union` (rels `deleteMany` old)
 
 -- | Like `Ix.deleteIx`, but works for multiple items
 deleteIxMulti ::